Fix failing tests
1. Refactor tests
2. Fix remaining tests
3. Fix Instrumented tests
4. Code cleanup
