+++ library/src/test/java/io/ona/kujaku/notifications/KujakuNotificationImplClass2.java
@@ -1,0 +1,8 @@
+package io.ona.kujaku.notifications;
+
+/**
+ * Created by Ephraim Kigamba - ekigamba@ona.io on 17/01/2018.
+ */
+
+public class KujakuNotificationImplClass2 extends KujakuNotification {
+}
